// TravellerCreator/ClientApp/src/app/models/sector-creator/hex/system/planet/trade-code.ts
export enum TradeCode {
  DataRepository = "Ab",
  AncientSite = "An",
  Agricultural = "Ag",
  AsteroidBelt = "As",
  Barren = "Ba",
  Cold = "Co",
  Capitol = "Cx",
  Desert = "De",
  Dieback = "Di",
  FluidOceans = "Fl",
  Frozen = "Fr",
  Garden = "Ga",
  Hellworld = "He",
  HighPopulation = "Hi",
  Hot = "Ho",
  HighTech = "Ht",
  IceCapped = "Ic",
  Industrial = "In",
  TidallyLocked = "Lk",
  LowPopulation = "Lo",
  LowTech = "Lt",
  MilitaryRule = "Mr",
  NonAgricultural = "Na",
  NonIndustrial = "Ni",
  Ocean = "Oc",
  PrisonCamp = "Px",
  Poor = "Po",
  Reserve = "Re",
  ResearchStation = "Rs",
  Rich = "Ri",
  Satellite = "Sa",
  Sterile = "St",
  Tropical = "Tr",
  Tundra = "Tu",
  TwilightZone = "Tz",
  WaterWorld = "Wa",
  Vacuum = "Va",
  Zoo = "Zo",
}